+++ Scripts/Complete-Firewall.ps1
@@ -28,7 +28,7 @@ SOFTWARE.
 
 <#PSScriptInfo
 
-.VERSION 0.9.1
+.VERSION 0.10.0
 
 .GUID cd8db1a8-63cd-443d-bdab-baf91b787fe9