+++ Scripts/Utility/Initialize-Development.ps1
@@ -27,7 +27,8 @@ SOFTWARE.
 #>
 
 <#PSScriptInfo
-.VERSION 0.9.1
+
+.VERSION 0.10.0
 
 .GUID 0390fb1d-bf61-441b-9746-63a34be384bb